[feat] SelectedImage 백그라운드 및 앨범 커버 공용 컴포넌트 제작
